+++ pre-gulp/assets/js/logoutbutton.js
@@ -1,0 +1,5 @@
+const logoutBtn = document.querySelector('.logout-button');
+
+logoutBtn.addEventListener('click', () => {
+    logoutBtn.classList.add('.logout-icon')
+})
