+++ pre-gulp/assets/js/start.js
@@ -1,0 +1,9 @@
+const signIn = document.getElementById('secondary');
+const section = document.querySelector('.options');
+const signUp = document.querySelector('.signup');
+
+signIn.addEventListener('click', () => {
+    section.classList.add('move');
+    signUp.classList.add('move-signup');
+    console.log("nu moet de section moven");
+})
