+++ public/js/logoutbutton.js
@@ -1,5 +1,2 @@
-const logoutBtn = document.querySelector('.logout-button');
-
-logoutBtn.addEventListener('click', () => {
-    logoutBtn.classList.add('.logout-icon')
-})
+const logoutBtn=document.querySelector(".logout-button");logoutBtn.addEventListener("click",(()=>{logoutBtn.classList.add(".logout-icon")}));
+//# sourceMappingURL=logoutbutton.js.map
