+++ public/js/start.js
@@ -1,9 +1,2 @@
-const signIn = document.getElementById('secondary');
-const section = document.querySelector('.options');
-const signUp = document.querySelector('.signup');
-
-signIn.addEventListener('click', () => {
-    section.classList.add('move');
-    signUp.classList.add('move-signup');
-    console.log("nu moet de section moven");
-})
+const signIn=document.getElementById("secondary"),section=document.querySelector(".options"),signUp=document.querySelector(".signup");signIn.addEventListener("click",(()=>{section.classList.add("move"),signUp.classList.add("move-signup"),console.log("nu moet de section moven")}));
+//# sourceMappingURL=start.js.map
