+++ public/js/userindex.js
@@ -1,102 +1,2 @@
-
-const filterBtn = document.querySelector('.filter-button');
-const sectionFilter = document.querySelector('.filter-container');
-const filterForm = document.querySelector('.filter-form');
-const userContainer = document.querySelector('.user-container');
-const users = document.querySelectorAll('.user');
-const userDetails = document.querySelector('.user-details');
-const userIndex = document.querySelector('.index');
-const container = document.querySelector('.container');
-
-console.log(container);
-
-const displayUsers = (data) =>{
-
-    userContainer.innerHTML = "";
-    data.forEach(match => {
-        const user = match.user;
-        let html = `<div class="user">
-                <div class="user-img">
-                    <img src="/public/uploads/${user.picture}" alt="">
-                </div>
-                <span>${user.firstname}</span>
-             </div>`
-
-        userContainer.insertAdjacentHTML('beforeend', html);
-
-    });
-}  
-
-const displaySingleUser = (data) => {
-    console.log(data);
-
-    let details = 
-    `    <section class="user-details">
-            <h2>${data.firstname}</h2>
-            <h2>${data.lastname}</h2>
-            <h2>${data.age}</h2>
-
-            <form action="/like" method="post">
-
-                <input type="hidden" name="id" value="${data._id}">
-                <input type="submit" value="like">
-            </form>
-        </section>`
-    userIndex.insertAdjacentHTML('beforeend', details);
-    closePopUp.classList.add('button-appear');
-}
-
-const filterUsers = (e) =>{
-    e.preventDefault();
-
-    const form = new FormData(e.target);
-    const goals = form.get("goals");
-
-    console.log(goals);
-
-    axios.post('/users/filter', {goals:goals})
-    .then(function (response) {
-        // handle success
-        console.log(response.data);
-        displayUsers(response.data);
-
-        sectionFilter.classList.remove('filter-up');
-    })
-    .catch(function (error) {
-        // handle error
-        console.log(error);
-    })
-    .then(function () {
-        // always executed
-    });
-}
-
-
-filterBtn.addEventListener('click', () => {
-    sectionFilter.classList.add('filter-up')
-})
-
-filterForm.addEventListener('submit', filterUsers);
-
-users.forEach(user => {
-    user.addEventListener('click', () =>{
-
-        const id = user.dataset.id;
-        axios.post('/users/getuser', {id:id})
-        .then(function (response) {
-            // handle success
-            displaySingleUser(response.data);
-        })
-        .catch(function (error) {
-            // handle error
-            console.log(error);
-        })
-    })
-})
-
-container.addEventListener('click', () => {
-
-    console.log(userDetails);
-    userDetails.classList.add('disappear');
-    console.log('nu moet iets geks gebeuren');
-})
+const filterBtn=document.querySelector(".filter-button"),sectionFilter=document.querySelector(".filter-container"),filterForm=document.querySelector(".filter-form"),userContainer=document.querySelector(".user-container"),users=document.querySelectorAll(".user"),userDetails=document.querySelector(".user-details"),userIndex=document.querySelector(".index"),container=document.querySelector(".container");console.log(container);const displayUsers=e=>{userContainer.innerHTML="",e.forEach((e=>{const n=e.user;let t=`<div class="user">\n                <div class="user-img">\n                    <img src="/public/uploads/${n.picture}" alt="">\n                </div>\n                <span>${n.firstname}</span>\n             </div>`;userContainer.insertAdjacentHTML("beforeend",t)}))},displaySingleUser=e=>{console.log(e);let n=`    <section class="user-details">\n            <h2>${e.firstname}</h2>\n            <h2>${e.lastname}</h2>\n            <h2>${e.age}</h2>\n\n            <form action="/like" method="post">\n\n                <input type="hidden" name="id" value="${e._id}">\n                <input type="submit" value="like">\n            </form>\n        </section>`;userIndex.insertAdjacentHTML("beforeend",n),closePopUp.classList.add("button-appear")},filterUsers=e=>{e.preventDefault();const n=new FormData(e.target).get("goals");console.log(n),axios.post("/users/filter",{goals:n}).then((function(e){var n;console.log(e.data),n=e.data,userContainer.innerHTML="",n.forEach((e=>{const n=e.user;let t=`<div class="user">\n                <div class="user-img">\n                    <img src="/public/uploads/${n.picture}" alt="">\n                </div>\n                <span>${n.firstname}</span>\n             </div>`;userContainer.insertAdjacentHTML("beforeend",t)})),sectionFilter.classList.remove("filter-up")})).catch((function(e){console.log(e)})).then((function(){}))};filterBtn.addEventListener("click",(()=>{sectionFilter.classList.add("filter-up")})),filterForm.addEventListener("submit",filterUsers),users.forEach((e=>{e.addEventListener("click",(()=>{const n=e.dataset.id;axios.post("/users/getuser",{id:n}).then((function(e){displaySingleUser(e.data)})).catch((function(e){console.log(e)}))}))})),container.addEventListener("click",(()=>{console.log(userDetails),userDetails.classList.add("disappear"),console.log("nu moet iets geks gebeuren")}));
+//# sourceMappingURL=userindex.js.map
