--- public/js/userindex.test.js
@@ -1,6 +1,0 @@
-const { TestWatcher } = require('@jest/core');
-const singleUser = require('./userindex');
-
-test('details of user is displayed', () => {
-    expect(singleUser) // geen idee....
-})
